\name{sampleClassifierData-package}
\alias{sampleClassifierData-package}
\alias{sampleClassifierData}


\docType{package}
\title{
\packageTitle{sampleClassifierData}
}
\description{
\packageDescription{sampleClassifierData}
}
\details{

This package contains a collection of publicly available microarray and RNA-seq datasets that have been pre-processed for use with the R-package 'sampleClassifier'.

The dataset \code{se_rnaseq_refmat} contains 71 samples from 24 tissues, derived from Fagerberg et al. (2014). Each tissue is represented by 3 replicates (except ovary which is represented by 2 replicates).

The dataset \code{se_micro_refmat} contains 78 samples from 26 tissues, derived from Roth et al. (2006). Each tissue is represented by 3 replicates.

The dataset \code{se_rnaseq_testmat} contains 12 samples from the Illumina Body Map 2.0 data.

The dataset \code{se_micro_testmat} contains 16 samples from Ge et al. (2005).

See the package vignette for details on the pre-processing of the data.
}
\author{
\packageAuthor{sampleClassifierData}

Maintainer: \packageMaintainer{sampleClassifierData}
}

\references{

Fagerberg, L., Hallstroem, B.M., Oksvold, P., Kampf, C., Djureinovic, D., Odeberg, J., Habuka, M., Tahmasebpoor, S., Danielsson, A., Edlund, K., Asplund, A., Sjoestedt, E., Lundberg, E., Szigyarto, C.A.-K., Skogs, M., Takanen, J.O., Berling, H., Tegel, H., Mulder, J., Nilsson, P., Schwenk, J.M., Lindskog, C., Danielsson, F., Mardinoglu, A., Sivertsson, A., von Feilitzen, K., Forsberg, M., Zwahlen, M., Olsson, I., Navani, S., Huss, M., Nielsen, J., Ponten, F., Uhlen, M.: Analysis of the human tissue-specific expression by genome-wide integration of transcriptomics and antibody-based proteomics. Molecular & cellular proteomics : MCP 13(2), 397-406 (2014).

Roth, R.B., Hevezi, P., Lee, J., Willhite, D., Lechner, S.M., Foster, A.C., Zlotnik, A.: Gene expression analyses reveal molecular relationships among 20 regions of the human CNS. Neurogenetics 7(2), 67-80 (2006). doi:10.1007/s10048-006-0032-6

The Illumina Body Map 2.0 data (https://www.ebi.ac.uk/arrayexpress/experiments/E-MTAB-513/).

Ge X, Yamamoto S, Tsutsumi S, Midorikawa Y, Ihara S, et al. Interpreting expression profiles of cancers by genome-wide survey of breadth of expression in normal tissues. Genomics 2005; 86:127-141. doi: 10.1016/j.ygeno.2005.04.008. pmid:15950434

}
