\name{se_micro_testmat}
\alias{se_micro_testmat}
\docType{data}
\title{
  microarray gene expression data set
}
\description{
  Microarray gene expression data set with 16 samples from Ge et al. (2005).
}
\usage{data(se_micro_testmat)}
\format{
  A \code{\link{matrix}} with 22283 probesets and 16 samples.
}
\value{
   microarray data matrix as SummarizedExperiment

}

\examples{
data(se_micro_testmat)

}
\keyword{dataset}
\references{
Ge X, Yamamoto S, Tsutsumi S, Midorikawa Y, Ihara S, et al. Interpreting expression profiles of cancers by genome-wide survey of breadth of expression in normal tissues. Genomics 2005; 86:127-141. doi: 10.1016/j.ygeno.2005.04.008. pmid:15950434
}