\name{se_rnaseq_testmat}
\alias{se_rnaseq_testmat}
\docType{data}
\title{
  RNA-seq gene expression data set
}
\description{
  RNA-seq gene expression data set with 12 samples from the Illumina Body Map 2.0 data.
}
\usage{data(se_rnaseq_testmat)}
\format{
  A \code{\link{matrix}} with 43819 genes and 12 samples.
}
\value{
   RNA-seq data matrix as SummarizedExperiment

}

\examples{
data(se_rnaseq_testmat)

}
\keyword{dataset}
\references{
The Illumina Body Map 2.0 data (https://www.ebi.ac.uk/arrayexpress/experiments/E-MTAB-513/).
}